backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==2.3.1
langchain==0.1.0
langchain-openai==0.0.5
openai==1.12.0
python-multipart==0.0.6
geopy
selenium
beautifulsoup4
webdriver-manager
googlemaps
pydantic